requirements.txt
asgiref==3.7.2
autopep8==2.0.2
darkdetect==0.8.0
distlib==0.3.7
Django==4.2.3
django-debug-toolbar==4.2.0
django-filter==23.3
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
filelock==3.12.2
fontawesomefree>=6.4.2
Markdown==3.4.4
#numpy==1.25.0
phonenumbers>=8.13.19
pillow>=10.0.1
platformdirs==3.9.1
psycopg==3.1.10
psycopg-binary>=3.1.12
psycopg-pool==3.1.7
pycodestyle==2.11.0
pytz==2023.3
reportlab==4.0.4
sqlparse==0.4.4
style==1.1.0
typing_extensions==4.7.1
tzdata==2023.3
update==0.0.1
virtualenv==20.24.2
django-autocomplete-light>=3.9.7

